+++ src/allotropy/allotrope/schema_parser/generate_schemas.py
@@ -96,8 +96,8 @@ def _generate_schema(model_path: Path, schema_path: Path) -> None:
         input_file_type=InputFileType.JsonSchema,
         # Specify base_class as empty when using dataclass
         base_class="",
-        target_python_version=PythonVersion.PY_39,
-        use_union_operator=False,
+        target_python_version=PythonVersion.PY_310,
+        use_union_operator=True,
     )
     # Import classes from shared files, remove unused classes, format.
     modify_file(str(model_path), str(schema_path))